refactor: Rename serial port QEMU driver path and update includes
