+++ alkos/kernel/arch/x86_64/boot/lib/debug_terminal.hpp
@@ -4,7 +4,7 @@
 /* internal includes */
 #include <assert.h>
 #include <autogen/feature_flags.h>
-#include <drivers/serial_port_qemu/serial_qemu.hpp>
+#include <drivers/serial/qemu.hpp>
 #include <extensions/defines.hpp>
 
 WRAP_CALL void DebugTerminalWrite(const char *const buffer)
