+++ alkos/kernel/arch/x86_64/boot/lib/drivers/serial/qemu.hpp
