+++ alkos/kernel/arch/x86_64/boot/lib/libc_backend.cpp
@@ -4,7 +4,7 @@
 #include <autogen/feature_flags.h> 
 #include <assert.h>
 #include <arch_utils.hpp> 
-#include <drivers/serial_port_qemu/serial_qemu.hpp>
+#include <drivers/serial/qemu.hpp>
 
 void __platform_panic(const char* msg)
 {
